requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
distro==1.9.0
Flask==3.1.2
flask-cors==6.0.1
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperon==0.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.15.0
Werkzeug==3.1.3
